feat: Chain Middleware to Create a Time Server
In the route app.get('/now', ...) chain a
middleware function and the final handler. In the
middleware function you should add the current
time to the request object in the req.time key.
You can use new Date().toString(). In the handler,
respond with a JSON object, taking the structure
{time: req.time}.