feat: Use body-parser to Parse POST Requests
body-parser has already been installed and is in
your project's package.json file. require it at
the top of the myApp.js file and store it in a
variable named bodyParser. The middleware to
handle URL encoded data is returned by
bodyParser.urlencoded({extended: false}).
Pass the function returned by the previous
method call to app.use(). As usual, the middleware
must be mounted before all the routes that depend
on it.